api/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.20
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.3
SQLAlchemy==2.0.44
psycopg2-binary==2.9.10
starlette==0.47.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.35.0
watchfiles==1.1.1
websockets==16.0
minio==7.2.18


